requirements.txt
pytest
request
allure-pytest
